refactor: :pencil2: Fix typo for locationComponent folder
