+++ src/components/locationComponent/LocationCard.vue
